Frontend/requirements.txt
flask
requests
python-dotenv
stripe
openai>=1.0.0
